added in-game text labels and information
- added accounting score, current level and goal to the next level to
the Model class
- added MainMesh::buildTextMesh function that build text string meshes
- added MainMesh::buildSidePanel function build interface panel meshes
- moved texture loading from View to MainMesh class
- fixed MainMesh shader dithering artefacts on dark colors
- removed deprecated Background class
- added font file
- rebuilt all executables

+++ src/Figure.cpp
@@ -1,3 +1,5 @@
+#include "static_headers.h"
+
 #include "Figure.h"
 
 int Figure::nextId = 1;
@@ -63,7 +65,7 @@ void Figure::buildFigure(Type type)
     assert(0);
   }
 
-  assert(strlen(cdata) == dim * dim);
+  assert((int)strlen(cdata) == dim * dim);
 
   cells.clear();
